+++ contracts/production/libraries/CosmicSignatureConstants.sol
@@ -29,8 +29,7 @@ library CosmicSignatureConstants {
 	/// This bit is to be set in `ICosmicSignatureGameStorage.RoundStats.flags`.
 	/// [/Comment-202610069]
 	/// It indicates that the same individual bid earned both Endurance Champion and Chrono-Warrior titles.
-	/// @dev todo-0 Revisit this constant type.
-	uint8 internal constant SAME_BID_EARNED_ENDURANCE_CHAMPION_AND_CHRONO_WARRIOR_TITLES = uint8(1 << 2);
+	uint256 internal constant SAME_BID_EARNED_ENDURANCE_CHAMPION_AND_CHRONO_WARRIOR_TITLES = 1 << 2;
 
 	// #endregion
 	// #region Bidding
